Consolidate provider and rule command code into one file each
provider.go, provider_add.go, and provider_cmd.go split the provider
command's Kong types and business logic across three files for no
reason beyond incremental history; merges them into provider_cmd.go.
Same for rule: export.go/import.go only existed to back `rule
export`/`rule import` (nothing else called them), so they're folded
into rule_cmd.go alongside the other rule subcommands.

Also merges the paired test files (provider_cli_test.go into
provider_cmd_test.go, rule_ci_test.go into rule_cmd_test.go) and drops
withControlledStdin, which only existed to feed `y\n` to runAdd's old
interactive-confirm path — dead now that provider add is pure CI.

provider_transfer.go is left separate since it's also used by the GUI
service, not just the provider CLI command.
